refactor: centralize theme toggle
